# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Local business discovery engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/review-queue:
    get:
      operationId: getReviewQueue
      tags: [businesses]
      summary: List businesses that should be manually reviewed next
      parameters:
        - name: categorySlug
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: targetMarket
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Prioritized review queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewQueueItem"
...
  /businesses/{id}/sources:
    get:
      operationId: getBusinessSources
      tags: [businesses]
      summary: List tracked sources for one business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sources for one business
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessSource"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /research/views:
    get:
      operationId: getResearchViews
      tags: [research]
      summary: List saved research views
      responses:
        "200":
          description: Saved views
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResearchView"
    post:
      operationId: createResearchView
      tags: [research]
      summary: Create a saved research view
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchViewPayload"
      responses:
        "201":
          description: Saved view created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchView"
        "400":
          description: Invalid saved view request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /research/views/{id}:
    patch:
      operationId: updateResearchView
      tags: [research]
      summary: Update a saved research view
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchViewPayload"
      responses:
        "200":
          description: Saved view updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchView"
        "404":
          description: Saved view not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteResearchView
      tags: [research]
      summary: Delete a saved research view
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Saved view deleted
...
  /export/businesses.csv:
    get:
      operationId: exportBusinessesCsv
      tags: [export]
      summary: Export businesses as CSV
      parameters:
        - name: categorySlug
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: targetMarket
          in: query
          schema:
            type: string
        - name: readyForOutreach
          in: query
          schema:
            type: boolean
        - name: reviewRequired
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
  /export/business-research-summary.csv:
    get:
      operationId: exportBusinessResearchSummaryCsv
      tags: [export]
      summary: Export the condensed research summary CSV
      parameters:
        - name: categorySlug
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: targetMarket
          in: query
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
...
components:
  schemas:
...
    Business:
      type: object
      properties:
        id:
          type: integer
        categorySlug:
          type: string
        name:
          type: string
        slug:
          type: string
        latitude:
          type: string
        longitude:
          type: string
        addressLine:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        osmId:
          type: ["string", "null"]
        osmType:
          type: ["string", "null"]
        googlePlaceId:
          type: ["string", "null"]
        googleMapsUrl:
          type: ["string", "null"]
        rating:
          type: ["string", "null"]
        userRatingsTotal:
          type: ["integer", "null"]
        hasWebsite:
          type: boolean
        hasPhone:
          type: boolean
        enrichmentStatus:
          type: string
        targetMarket:
          type: ["string", "null"]
        engineType:
          type: ["string", "null"]
        targetType:
          type: ["string", "null"]
        targetCluster:
          type: ["string", "null"]
        avatarType:
          type: ["string", "null"]
        discoveryStatus:
          type: ["string", "null"]
        qualificationStatus:
          type: ["string", "null"]
        contactabilityStatus:
          type: ["string", "null"]
        rankingStatus:
          type: ["string", "null"]
        sourceHealth:
          type: ["string", "null"]
        contactReadiness:
          type: ["string", "null"]
        relevanceScore:
          type: ["integer", "null"]
        contactabilityScore:
          type: ["integer", "null"]
        confidenceScore:
          type: ["integer", "null"]
        freshnessScore:
          type: ["integer", "null"]
        priorityScore:
          type: ["integer", "null"]
        researchScore:
          type: ["integer", "null"]
        economicValueScore:
          type: ["integer", "null"]
        strategicValueScore:
          type: ["integer", "null"]
        referralValueScore:
          type: ["integer", "null"]
        prestigeValueScore:
          type: ["integer", "null"]
        continuityRevenuePotential:
          type: ["integer", "null"]
        offerFitScore:
          type: ["integer", "null"]
        relationshipPathScore:
          type: ["integer", "null"]
        actionabilityScore:
          type: ["integer", "null"]
        officialSourceCount:
          type: ["integer", "null"]
        successfulSourceCount:
          type: ["integer", "null"]
        failedSourceCount:
          type: ["integer", "null"]
        primarySourceId:
          type: ["integer", "null"]
        primaryContactCandidateId:
          type: ["integer", "null"]
        bestOfferId:
          type: ["integer", "null"]
        bestNarrativeId:
          type: ["integer", "null"]
        secondaryNarrativeId:
          type: ["integer", "null"]
        bestCredibilityAssetId:
          type: ["integer", "null"]
        bestCaseStudyId:
          type: ["integer", "null"]
        proofAngle:
          type: ["string", "null"]
        riskReductionReason:
          type: ["string", "null"]
        toneOfApproach:
          type: ["string", "null"]
        recommendedPitchAngle:
          type: ["string", "null"]
        nextBestContactWindow:
          type: ["string", "null"]
        accountTier:
          type: ["string", "null"]
        seasonalityFit:
          type: ["string", "null"]
        warmPathExists:
          type: boolean
        readyForRelationship:
          type: boolean
        readyForInstitutionalPitch:
          type: boolean
        prestigeWatchlist:
          type: boolean
        cultivationRequired:
          type: boolean
        readyForOutreach:
          type: boolean
        reviewRequired:
          type: boolean
        reviewReason:
          type: ["string", "null"]
        topGap:
          type: ["string", "null"]
        recommendedNextStep:
          type: ["string", "null"]
        lastResearchAt:
          type: ["string", "null"]
        nextResearchAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - categorySlug
        - name
        - slug
        - latitude
        - longitude
        - hasWebsite
        - hasPhone
        - enrichmentStatus
        - warmPathExists
        - readyForRelationship
        - readyForInstitutionalPitch
        - prestigeWatchlist
        - cultivationRequired
        - readyForOutreach
        - reviewRequired
        - createdAt
        - updatedAt
...
    BusinessSource:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        sourceType:
          type: string
        sourceUrl:
          type: string
        sourceDomain:
          type: ["string", "null"]
        discoveredVia:
          type: ["string", "null"]
        fetchStatus:
          type: string
        lastFetchedAt:
          type: ["string", "null"]
        contentHash:
          type: ["string", "null"]
        httpStatus:
          type: ["integer", "null"]
        isOfficial:
          type: boolean
        sourcePriority:
          type: integer
        usefulnessScore:
          type: ["integer", "null"]
        nextFetchAt:
          type: ["string", "null"]
        freshnessStatus:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - businessId
        - sourceType
        - sourceUrl
        - fetchStatus
        - isOfficial
        - sourcePriority
        - freshnessStatus
        - createdAt
        - updatedAt
    ReviewQueueItem:
      type: object
      properties:
        business:
          $ref: "#/components/schemas/Business"
        reasons:
          type: array
          items:
            type: string
        priorityScore:
          type: integer
        sourceCount:
          type: integer
        officialSourceCount:
          type: integer
        failedSourceCount:
          type: integer
      required:
        - business
        - reasons
        - priorityScore
        - sourceCount
        - officialSourceCount
        - failedSourceCount
...
    ResearchView:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        scope:
          type: string
        isDefault:
          type: boolean
        filtersJson:
          type: object
          additionalProperties: true
        sortJson:
          type: object
          additionalProperties: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, scope, isDefault, filtersJson, sortJson]
    ResearchViewPayload:
      type: object
      properties:
        name:
          type: string
        scope:
          type: string
        isDefault:
          type: boolean
        filtersJson:
          type: object
          additionalProperties: true
        sortJson:
          type: object
          additionalProperties: true
      required: [name]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]